requirements.txt
requests
bs4
argparse
tqdm
Pillow
opencv-python
imagehash